+++ src/proto41.h
@@ -9,7 +9,7 @@
 
 #include "sixxsd.h"
 
-VOID proto41_in(const IPADDRESS *src, uint8_t *packet, const uint32_t len);
+VOID proto41_in(const IPADDRESS *src, uint8_t *packet, const uint16_t len);
 VOID proto41_out(const uint16_t in_tid, const uint16_t out_tid, const uint8_t *packet, const uint16_t len, BOOL is_response);
 
 #endif
